# 2 Optimizations when verifying stake