--- a/src/main.cpp
+++ b/src/main.cpp
@@ -1977,6 +1977,9 @@ bool CBlock::AddToBlockIndex(unsigned int nFile, unsigned int nBlockPos)
 
     // Add to mapBlockIndex
     map<uint256, CBlockIndex*>::iterator mi = mapBlockIndex.insert(make_pair(hash, pindexNew)).first;
+    // Link pnextWithStakeModifier chains so we can skip around faster
+    pindexNew->linkPnextWithStakeModifier();
+
     if (pindexNew->IsProofOfStake())
         setStakeSeen.insert(make_pair(pindexNew->prevoutStake, pindexNew->nStakeTime));
     pindexNew->phashBlock = &((*mi).first);