--- a/src/main.h
+++ b/src/main.h
@@ -945,6 +945,9 @@ public:
     mutable int nDoS;
     bool DoS(int nDoSIn, bool fIn) const { nDoS += nDoSIn; return fIn; }
 
+    // CheckBlock only compute once
+    mutable bool bComputedCheckBlock, bCheckBlockReturn;
+
     CBlock()
     {
         SetNull();
@@ -970,6 +973,8 @@ public:
         vchBlockSig.clear();
         vMerkleTree.clear();
         nDoS = 0;
+        bComputedCheckBlock = false;
+        bCheckBlockReturn = false;
     }
 
     // ppcoin: entropy bit for stake modifier if chosen by modifier