This sits on top of Pull request #25, #26 and #27  

- Cached hashes calculated when computing stake chains
- Allow jumping CBlockIndex directly to next stake modifier block. This is a potentially dangerous change, need to test what happens in the scenario of finding a new best chain and rewritting the chain. Remove commit if it is a problem. 5c7cafa

Tested by looking at LoadExternalBlockFile and every 500 blocks printed (total elapsed time/ nloaded) in ms
This change effects the rate of scaling and helps more on larger nLoaded

nLoaded | Avg Before | Avg #25 | Avg #26 | Avg #27 | Avg this
---|---|---|---|---|---
500 | 4.98ms | 2.54ms | 1.80ms | 1.77ms | 1.75ms
20000 | 5.96ms | 3.38ms | 2.79ms | 2.59ms | 2.45ms
100000 | 7.97ms | 5.26ms | 4.75ms | 3.76ms | 3.00ms
200000 | - | - | - | - | 3.23ms