@pol081 2017-06-11
The average was 4.12ms out at 2 million nLoaded